% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg.svy.bubble.R
\name{gg.svy.bubble}
\alias{gg.svy.bubble}
\title{Takes x-values, y-values, bubble-sizes and color category to make bubble plot from svystat objects.}
\usage{
gg.svy.bubble(xtab, ytab, stab, cols, colors, base_size = 11, alpha = 0.3,
  discrete = FALSE, x.var = NULL, y.var = NULL)
}
\arguments{
\item{xtab}{X-values calculated with svyby().}

\item{ytab}{Y-values calculated with svyby().}

\item{stab}{Size values calculated with svymean().}

\item{cols}{A vector of values to color-by of length equal to number of categories in svyby.}

\item{colors}{A color palette.}

\item{base_size}{Default font size.}

\item{alpha}{Transparency for bubble fill. Default is 0.3.}

\item{discrete}{If TRUE replace numbers on x and y-axes with factor levels. Must also provide x.var and y.var.}

\item{x.var}{A factor for x-axis, e.g. data$x.var.}

\item{y.var}{A factor for x-axis, e.g. data$x.var.}
}
\value{
A ggplot object.
}
\description{
Takes x-values, y-values, bubble-sizes and color category to make bubble plot from svystat objects.
}
