% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg.svy.likert.R
\name{gg.svy.likert}
\alias{gg.svy.likert}
\title{Generates likert plot based on svyby stat}
\usage{
gg.svy.likert(tab, order = FALSE, colors = NULL, line = 8,
  legend.position = "bottom", legend.name = "", wrap.levels = FALSE,
  level.width = 15, base_size = 11)
}
\arguments{
\item{tab}{The \code{svystat} object from the function \code{svymean}.}

\item{order}{Plot bars in decending order? Do not use if factor is already in specific order, e.g. survey waves.}

\item{line}{Line width. Default is 6.}

\item{legend.position}{Legend position.}

\item{legend.name}{Name for legend.}

\item{color}{A color list generated with \code{\link{gg.svy.colors}} (no quotes) or Hex color codes (in quotes).}
}
\value{
A ggplot object.
}
\description{
Creates a likert-type plot.
}
