% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barcharts.R
\name{ggbarcrosstabs3d_svy}
\alias{ggbarcrosstabs3d_svy}
\title{Crosstabs of Three Variables Using svy.design object}
\usage{
ggbarcrosstabs3d_svy(surveyobj, x, y, z)
}
\arguments{
\item{surveyobj}{svy.design obj}

\item{x}{bar chart variable}

\item{y}{crosstab variable 1 (horizontal facets)}

\item{z}{crosstab variable 2 (vertical facets)}
}
\value{

}
\description{
In ggsurvey you specify both the plotting variables and weights in plain text with no quotes.
}
\examples{
library(survey)
data(api)
dstrat<-svydesign(id=~1,strata=~stype, weights=~pw, data=apistrat, fpc=~fpc)
ggbarcrosstabs3d_svy(dstrat, stype, yr.rnd, awards)
data(nhanes)
design <- svydesign(id=~SDMVPSU, strata=~SDMVSTRA, weights=~WTMEC2YR, nest=TRUE,data=nhanes)+ylab("Proportion")
ggbarcrosstabs3d_svy(design, race, agecat, RIAGENDR)
}
