% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barcharts.R
\name{ggbarweight_svy}
\alias{ggbarweight_svy}
\title{Bar Chart from svydesign objects}
\usage{
ggbarweight_svy(surveyobj, x)
}
\arguments{
\item{surveyobj}{svydesign}

\item{x}{variable to plot}
}
\value{
ggplot object
}
\description{
Bar Chart from svydesign objects
}
\examples{
library(survey)
data(api)
dstrat<-svydesign(id=~1,strata=~stype, weights=~pw, data=apistrat, fpc=~fpc)
ggbarweight_svy(dstrat, stype)+ylab("Proportion")
data(nhanes)
design <- svydesign(id=~SDMVPSU, strata=~SDMVSTRA, weights=~WTMEC2YR, nest=TRUE,data=nhanes)+ylab("Proportion")
ggbarweight_svy(design, race)
}
