% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplots.R
\name{ggboxweight}
\alias{ggboxweight}
\title{Weighted Box Plot of One Variable}
\usage{
ggboxweight(df, x, weights)
}
\arguments{
\item{df}{data frame}

\item{x}{first variable of interest}

\item{weights}{survey weights that sums to sample size}
}
\value{

}
\description{
Weighted Box Plot of One Variable
}
\examples{
library(survey)
data(api)
ggboxweight(apistrat, api00, pw)
data(election)
ggboxweight(election_pps, Bush, p)
}
