% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplots.R
\name{ggboxweight2d}
\alias{ggboxweight2d}
\title{Weighted Box Plot with a categorical variable}
\usage{
ggboxweight2d(df, x, y, weights)
}
\arguments{
\item{df}{data frame}

\item{x}{categorical variable of interest}

\item{y}{numeric variable of interest}

\item{weights}{survey weights that sums to sample size}
}
\value{

}
\description{
Weighted Box Plot with a categorical variable
}
