% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplots.R
\name{ggboxweight2d_svy}
\alias{ggboxweight2d_svy}
\title{Weighted Boxplot of a survey object with a categorical variable}
\usage{
ggboxweight2d_svy(surveyobj, x, y)
}
\arguments{
\item{surveyobj}{svy.design object}

\item{x}{variable to boxplot}

\item{y}{categorical variable}
}
\value{
ggplot object
}
\description{
In ggsurvey you specify both the plotting variables and weights in plain text with no quotes.
}
\examples{
library(survey)
data(api)
dstrat<-svydesign(id=~1,strata=~stype, weights=~pw, data=apistrat, fpc=~fpc)
ggboxweight2d_svy(dstrat, api00, stype)
}
