% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplots.R
\name{ggboxweight3d_svy}
\alias{ggboxweight3d_svy}
\title{Weighted Boxplot of svy.design object with two categorical variables}
\usage{
ggboxweight3d_svy(surveyobj, x, y, z)
}
\arguments{
\item{surveyobj}{svy.design}

\item{x}{variable to boxplot}

\item{y}{first categorical variable}

\item{z}{second categorical variable (for faceting)}
}
\value{

}
\description{
Weighted Boxplot of svy.design object with two categorical variables
}
\examples{
library(survey)
data(api)
dstrat<-svydesign(id=~1,strata=~stype, weights=~pw, data=apistrat, fpc=~fpc)
ggboxweight3d_svy(dstrat, api00, stype, awards)
}
