% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barcharts.R
\name{ggcrosstabs}
\alias{ggcrosstabs}
\title{Crosstabs of Two Variables}
\usage{
ggcrosstabs(df, x, y, weights)
}
\arguments{
\item{df}{survey data frame}

\item{x}{first variable of interest}

\item{y}{second variable of interest}

\item{weights}{survey weights that sum to sample size}
}
\value{

}
\description{
Crosstabs of Two Variables
}
