% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barcharts.R
\name{ggfreqweight}
\alias{ggfreqweight}
\title{Weighted Bar Charts}
\usage{
ggfreqweight(df, x, weights)
}
\arguments{
\item{df}{survey dataframe}

\item{x}{name of question of interest}

\item{weights}{survey weights that sums to sample size}
}
\value{
ggplot
}
\description{
Weighted Bar Charts
}
