% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hexplot.R
\name{gghexweight}
\alias{gghexweight}
\title{Weighted Hex Plot}
\usage{
gghexweight(df, x, y, weights)
}
\arguments{
\item{df}{data frame}

\item{x}{name of variable for x axis}

\item{y}{name of variable for y axis}

\item{weights}{name of weights variable}
}
\value{

}
\description{
In ggsurvey you specify both the plotting variables and weights in plain text with no quotes.
}
\examples{
data(api)
gghexweight(apistrat, api99, api00, pw)
}
