% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hexplot.R
\name{gghexweight2d}
\alias{gghexweight2d}
\title{Weighted Hex Plot with One Facet Variable}
\usage{
gghexweight2d(df, x, y, z, weights)
}
\arguments{
\item{df}{data frame}

\item{x}{name of variable for x axis}

\item{y}{name of variable for y axis}

\item{z}{faceting variable}

\item{weights}{desired binwidth, if NULL bins in geom_histogram defaults to 30}
}
\value{

}
\description{
Weighted Hex Plot with One Facet Variable
}
