% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hexplot.R
\name{gghexweight2d_svy}
\alias{gghexweight2d_svy}
\title{Weighted Hex Plot of svy.design with One Facet Variable}
\usage{
gghexweight2d_svy(surveyobj, x, y, z)
}
\arguments{
\item{surveyobj}{svy.design}

\item{x}{variable for x axis}

\item{y}{variable for y axis}

\item{z}{faceting variable}
}
\value{
ggplot
}
\description{
In ggsurvey you specify both the plotting variables and weights in plain text with no quotes.
}
\examples{
library(survey)
data(api)
dstrat<-svydesign(id=~1,strata=~stype, weights=~pw, data=apistrat, fpc=~fpc)
gghexweight2d_svy(dstrat, api99, api00, stype)
}
