% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hexplot.R
\name{gghexweight3d}
\alias{gghexweight3d}
\title{Weighted Box Plot with Two Facet Variables}
\usage{
gghexweight3d(df, x, y, a, b, weights)
}
\arguments{
\item{df}{data frame}

\item{x}{name of variable for x axis}

\item{y}{name of variable for y axis}

\item{a}{first faceting variable}

\item{b}{second faceting variable}

\item{weights}{name of weights variable}
}
\value{
ggplot object
}
\description{
In ggsurvey you specify both the plotting variables and weights in plain text with no quotes.
}
\examples{
data(api)
gghexweight3d(apistrat, api99, api00, stype, awards, pw)
}
