% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hexplot.R
\name{gghexweight3d_svy}
\alias{gghexweight3d_svy}
\title{Weighted Hex Plot of svy.design with Two Faceting Variables}
\usage{
gghexweight3d_svy(surveyobj, x, y, a, b)
}
\arguments{
\item{surveyobj}{svy.design}

\item{x}{variable for x axis}

\item{y}{variable for y axis}

\item{a}{horizontal facetting variable}

\item{b}{vertical facetting variable}
}
\value{

}
\description{
In ggsurvey you specify both the plotting variables and weights in plain text with no quotes.
}
\examples{
library(survey)
data(api)
dstrat<-svydesign(id=~1,strata=~stype, weights=~pw, data=apistrat, fpc=~fpc)
gghexweight3d_svy(dstrat, api99, api00, stype, awards)
}
