% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histograms.R
\name{gghistweight2d}
\alias{gghistweight2d}
\title{Weighted Histogram with One Facet
In ggsurvey you specify both the plotting variables and weights in plain text with no quotes.}
\usage{
gghistweight2d(df, x, y, weights, binwidth = NULL)
}
\arguments{
\item{df}{data frame}

\item{x}{first variable of interest}

\item{y}{categorical variable for faceting}

\item{weights}{survey weights that sum to sample size}

\item{binwidth}{desired binwidth, if NULL bins in geom_histogram defaults to 30}
}
\value{
ggplot object
}
\description{
Weighted Histogram with One Facet
In ggsurvey you specify both the plotting variables and weights in plain text with no quotes.
}
\examples{
library(survey)
data(api)
gghistweight2d(apistrat, api00, stype, pw)
gghistweight2d(apistrat, api00, stype, pw, binwidth = 10)
}
