% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histograms.R
\name{gghistweight_svy}
\alias{gghistweight_svy}
\title{Histogram of svgdesign object}
\usage{
gghistweight_svy(surveyobj, x, binwidth = NULL)
}
\arguments{
\item{surveyobj}{svy.design object}

\item{x}{variable to histogram}

\item{binwidth}{binwidth to pass to geom_hist}
}
\value{

}
\description{
In ggsurvey you specify both the plotting variables and weights in plain text with no quotes.
}
\examples{
library(survey)
data(api)
dstrat<-svydesign(id=~1,strata=~stype, weights=~pw, data=apistrat, fpc=~fpc)
gghistweight_svy(dstrat, api00)
gghistweight_svy(dstrat, api00, binwidth = 10)
}
