% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qol-data.R
\docType{data}
\name{qol}
\alias{qol}
\title{Subjective Quality of Life Survey Data}
\format{A data frame with 1594 rows and 21 columns.}
\usage{
data(qol)
}
\description{
Data from a survey of four Palestinian refugee camps in Lebanon.
}
\details{
\itemize{
  \item camp. Name of the refugee camp.
  \item gender. Gender of respondent.
  \item prs. Respondent is a Palestinian refugee from Syria.
  \item age. Age of respondent.
  \item i_posessions How important are the following to you? YOUR POSESSIONS.
  \item i_health. How important are the following to you? YOUR HEALTH.
  \item i_achievements. How important are the following to you? YOUR ACHIEVEMENTS.
  \item i_family. How important are the following to you? STRONG BONDS WITH FAMILY AND FRIENDS.
  \item i_safety. How important are the following to you? FEELING SAFE.
  \item i_outdoor. How important are the following to you? OUTDOOR ACTIVITIES WITH OTHER PEOPLE.
  \item s_posessions. How satisfied are you with each of these? YOUR POSESSIONS.
  \item s_health. How satisfied are you with each of these? YOUR HEALTH.
  \item s_achievements. How satisfied are you with each of these? YOUR ACHIEVEMENTS.
  \item s_family. How satisfied are you with each of these? STRONG BONDS WITH FAMILY AND FRIENDS.
  \item s_safety. How satisfied are you with each of these? FEELING SAFE.
  \item s_outdoor. How satisfied are you with each of these? OUTDOOR ACTIVITIES WITH OTHER PEOPLE.
  \item adult_weight. Probability weight.
  \item sqol_sm. SQOL satisfaction items weighted by importance. In standard measures (0--100).
  \item fpc. Stratum (camp) population sizes, for finite population correction.
}
}
\keyword{datasets}
