% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg.svyby.R
\name{gg.svyby}
\alias{gg.svyby}
\title{Generates plot based on svyby stat}
\usage{
gg.svyby(tab, basic = FALSE, color = NULL, percent = FALSE,
  error.bars = TRUE, outcome.factor = TRUE, legend.position = "right",
  legend.title = "Response", drop.level = NULL, error.bar.bounds = FALSE)
}
\arguments{
\item{tab}{The \code{svystat} object from the function \code{svymean}.}

\item{basic}{If TRUE, only a basic ggplot object is initialized.}

\item{color}{A color list generated with \code{\link{gg.svy.colors}} (no quotes) or a Hex color codes (in quotes).}

\item{percent}{If TRUE, multiplies all decimal values by 100.}

\item{error.bars}{If TRUE, adds error bars for 95\% confidence intervals.}
}
\value{
A ggplot object, with or without additional aesthetics, depending on \code{basic} parameter.
}
\description{
This function will either initialize a basic ggplot
or generate a barblot with additional aesthetics. The \code{color}
parameter generated with \code{\link{gg.svy.colors}} can be used
to specify a color mapping.
}

