% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barcharts.R
\name{ggbarcrosstabs}
\alias{ggbarcrosstabs}
\title{Crosstabs of Two Variables}
\usage{
ggbarcrosstabs(df, x, y, weights)
}
\arguments{
\item{df}{data frame of survey}

\item{x}{variable to bar chart}

\item{y}{faceting variable}

\item{weights}{survey weights that sum to sample size}
}
\value{
ggplot object that can be customized using ggplot2 functions including adding titles
}
\description{
In ggsurvey you specify both the plotting variables and weights in plain text with no quotes.
}
\examples{
library(survey)
data(api)
ggbarcrosstabs(apistrat, stype, yr.rnd, pw)
data(nhanes)
ggbarcrosstabs(nhanes, race, agecat, WTMEC2YR)
}
