% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barcharts.R
\name{ggbarcrosstabs3d}
\alias{ggbarcrosstabs3d}
\title{Crosstabs of Three Variables}
\usage{
ggbarcrosstabs3d(df, x, y, z, weights)
}
\arguments{
\item{df}{data frame}

\item{x}{bar chart variable}

\item{y}{crosstab variable 1 (horizontal facets)}

\item{z}{crosstab variable 2 (vertical facets)}

\item{weights}{survey weights that sum to sample size}
}
\value{

}
\description{
Crosstabs of Three Variables
}
\examples{
library(survey)
data(api)
ggbarcrosstabs3d(apistrat, stype, yr.rnd, awards, pw)
data(nhanes)
ggbarcrosstabs3d(nhanes, race, agecat, RIAGENDR, WTMEC2YR)
}
