% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barcharts.R
\name{ggbarcrosstabs3d}
\alias{ggbarcrosstabs3d}
\title{Crosstabs of Three Variable}
\usage{
ggbarcrosstabs3d(df, x, y, z, weights)
}
\arguments{
\item{df}{survey data frame}

\item{x}{first variable of interest}

\item{y}{second variable of interest}

\item{z}{third variable of interest}

\item{weights}{survey weights that sum to sample size}
}
\value{

}
\description{
Crosstabs of Three Variable
}
