% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barcharts.R
\name{ggbarweight_svy}
\alias{ggbarweight_svy}
\title{Bar Chart from svydesign objects}
\usage{
ggbarweight_svy(surveyobj, x)
}
\arguments{
\item{surveyobj}{svydesign}

\item{x}{variable to plot}
}
\value{
ggplot object
}
\description{
In ggsurvey you specify both the plotting variables and weights in plain text with no quotes.
}
