% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplots.R
\name{ggboxweight2d}
\alias{ggboxweight2d}
\title{Weighted Boxplot with a categorical variable}
\usage{
ggboxweight2d(df, x, y, weights)
}
\arguments{
\item{df}{data frame}

\item{x}{numeric variable of interest}

\item{y}{categorical variable of interest}

\item{weights}{survey weights that sums to sample size}
}
\value{

}
\description{
In ggsurvey you specify both the plotting variables and weights in plain text with no quotes.
}
\examples{
library(survey)
data(api)
ggboxweight2d(apistrat, api00, stype, pw)
}
