% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplots.R
\name{ggboxweight_svy}
\alias{ggboxweight_svy}
\title{Weighted Box Plot of svy.design object}
\usage{
ggboxweight_svy(surveyobj, x)
}
\arguments{
\item{surveyobj}{svy.design object}

\item{x}{variable to boxplot}
}
\value{
ggplot boxplot
}
\description{
Weighted Box Plot of svy.design object
}
\examples{
library(survey)
data(api)
dstrat<-svydesign(id=~1,strata=~stype, weights=~pw, data=apistrat, fpc=~fpc)
ggboxweight_svy(dstrat, api00)
}
