% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hexplot.R
\name{gghexweight}
\alias{gghexweight}
\title{Weighted Hex Plot}
\usage{
gghexweight(df, x, y, weights)
}
\arguments{
\item{df}{data frame}

\item{x}{name of variable for x axis}

\item{y}{name of variable for y axis}

\item{weights}{desired binwidth, if NULL bins in geom_histogram defaults to 30}
}
\value{

}
\description{
Weighted Hex Plot
}
\examples{
data(api)
gghexweight(apistrat, api99, api00, pw)
}
