% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hexplot.R
\name{gghexweight2d}
\alias{gghexweight2d}
\title{Weighted Hex Plot with One Facet Variable}
\usage{
gghexweight2d(df, x, y, z, weights)
}
\arguments{
\item{df}{data frame}

\item{x}{name of variable for x axis}

\item{y}{name of variable for y axis}

\item{z}{faceting categorical variable}

\item{weights}{name of weights variable}
}
\value{
ggplot object
}
\description{
In ggsurvey you specify both the plotting variables and weights in plain text with no quotes.
}
\examples{
data(api)
gghexweight2d(apistrat, api99, api00, stype, pw)
}
