% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hexplot.R
\name{gghexweight_svy}
\alias{gghexweight_svy}
\title{Weighted Hex Plot of Survey Design Object}
\usage{
gghexweight_svy(surveyobj, x, y)
}
\arguments{
\item{surveyobj}{svy.design}

\item{x}{name of variable for x axis}

\item{y}{name of variable for y axis}
}
\value{

}
\description{
Weighted Hex Plot of Survey Design Object
}
\examples{
library(survey)
data(api)
dstrat<-svydesign(id=~1,strata=~stype, weights=~pw, data=apistrat, fpc=~fpc)
gghexweight_svy(dstrat, api99, api00)
}
