% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histograms.R
\name{gghistweight}
\alias{gghistweight}
\title{Weighted Histograms}
\usage{
gghistweight(df, x, weights, binwidth = NULL)
}
\arguments{
\item{df}{data frame}

\item{x}{variable of interest}

\item{weights}{survey weights that sum to sample size}

\item{binwidth}{desired binwidth, if NULL bins in geom_histogram defaults to 30}
}
\value{

}
\description{
Weighted Histograms
}
